+++ contracts/mocks/MerkleProofWrapper.sol
@@ -12,4 +12,8 @@ contract MerkleProofWrapper {
     ) public pure returns (bool) {
         return MerkleProof.verify(proof, root, leaf);
     }
+
+    function processProof(bytes32[] memory proof, bytes32 leaf) public pure returns (bytes32) {
+        return MerkleProof.processProof(proof, leaf);
+    }
 }