+++ contracts/proxy/beacon/UpgradeableBeacon.sol
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: MIT
+// OpenZeppelin Contracts v4.3.2 (proxy/beacon/UpgradeableBeacon.sol)
 
 pragma solidity ^0.8.0;