+++ contracts/token/ERC20/ERC20.sol
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: MIT
+// OpenZeppelin Contracts v4.3.2 (token/ERC20/ERC20.sol)
 
 pragma solidity ^0.8.0;